Fixa merge och reMerge metoderna

+++ src/ghzclicker/Controller.java
@@ -66,10 +66,10 @@ public class Controller {
 	 * This dose so if hertz=1000, we will get 1Khz and 0 Hertz
 	 */
 	public void merging() {
-		double diff;
+		int diff;
 		for (int i = 0; i < hertz.size(); i++) {
 			if (hertz.get(i) >= 1000) {
-				diff = (hertz.get(i) / 1000);
+				diff = (int) (hertz.get(i) / 1000);
 				hertz.set(i, (hertz.get(i) - diff * 1000));
 				hertz.set(i + 1, (hertz.get(i + 1) + diff));
 			}
@@ -80,12 +80,12 @@ public class Controller {
 	 * This dose so if hertz gets under 0 we will take from KHz and give to hertz
 	 */
 	public void reMerge() {
-		double diff;
+		int diff;
 		for (int i = 0; i < hertz.size() - 1; i++) {
 			if (hertz.get(i) < 0) {
-				diff = Math.abs(hertz.get(i)) / 1000;
-				hertz.set(i + 1, (hertz.get(i + 1) - (1 + diff)));
-				hertz.set(i, (hertz.get(i) + diff * 1000 + 1000));
+				diff = (int) Math.abs(hertz.get(i)) / 1000;			
+				hertz.set(i+1, (hertz.get(i+1) - (1 + diff)));			
+				hertz.set(i, (hertz.get(i) + diff * 1000 + 1000));			
 			}
 		}
 	}
